fix: break recovery type dependency cycle
