+++ packages/api/src/types/earlyBufferRecovery.ts
@@ -1,0 +1,22 @@
+export type EarlyBufferRecoveryMethod = 'redis' | 'snapshot';
+export type EarlyBufferRecoveryOutcome = 'success' | 'failed';
+export type EarlyBufferRecoveryFailureReason =
+  | 'durable_state_missing'
+  | 'durable_frontier_gap'
+  | 'snapshot_missing'
+  | 'subscriber_never_attached'
+  | 'subscriber_disconnected'
+  | 'reconstruction_error'
+  | 'overflow_marker_persistence_failed';
+
+export interface EarlyBufferOverflowState {
+  id: string;
+  occurredAt: number;
+  durableEvents: number;
+  droppedEvents: number;
+  droppedBytes: number;
+  recoveryMethod?: EarlyBufferRecoveryMethod;
+  recoveryOutcome?: EarlyBufferRecoveryOutcome;
+  recoveryCompletedAt?: number;
+  recoveryFailureReason?: EarlyBufferRecoveryFailureReason;
+}
